feat(protocol): MP-BGP IPv6/Unicast — phase 2 of #14/#15 (#402)
Phase 2 of the IPv6 epic: MP_REACH_NLRI (attr 14) and MP_UNREACH_NLRI (attr 15) codec for IPv6/Unicast, capability advertisement and detection, session receive path for IPv6 routes. Route.NextHop widened to UInt128. 949/949 tests (9 new codec tests).

+++ BGPLite.Routing/Route.cs
@@ -10,11 +10,12 @@ public sealed class Route
 {
     // #15 phase 1: 128-bit prefix — IPv4 in the low 32 bits with IsIpv4 = true (the implicit
     // uint → UInt128 conversion in object initializers lands exactly there), IPv6 in the full
-    // 128 bits with IsIpv4 = false. NextHop stays IPv4 (uint) until the MP_REACH phase.
+    // 128 bits with IsIpv4 = false. #15 phase 2: NextHop is likewise 128-bit (IPv4 low bits);
+    // for MP_REACH routes it carries the peer's IPv6 next hop.
     public required UInt128 Prefix { get; init; }
     public bool IsIpv4 { get; init; } = true;
     public required byte PrefixLength { get; init; }
-    public required uint NextHop { get; init; }
+    public required UInt128 NextHop { get; init; }
     public IReadOnlyList<uint> AsPath { get; init; } = [];
     public IReadOnlyList<uint> Communities { get; init; } = [];
     /// <summary>BGP Large Communities (RFC 8092): triplets of (Global : Local1 : Local2).</summary>
